requirements.txt
uvicorn==0.14.0
fastapi==0.68.0
alembic==1.6.5
databases==0.3.2
psycopg2-binary==2.9.1
pytest==6.2.4
requests==2.26.0
python-dotenv==0.19.0
SQLAlchemy==1.3.24
asyncpg==0.23.0
email-validator==1.1.3
argon2-cffi==20.1.0
PyJWT==2.1.0
cryptography==3.4.7
aiohttp==3.6.2
pytest-mock==3.6.1
pandas==1.3.1
